+++ src/elpi_compiler.mli
@@ -3,6 +3,7 @@
 (* ------------------------------------------------------------------------- *)
 
 open Elpi_util
+open Elpi_data
 
 type flags = {
   defined_variables : StrSet.t;
@@ -18,27 +19,28 @@ type query
 val program_of_ast : flags:flags -> Elpi_ast.Program.t -> program
 val query_of_ast : program -> Elpi_ast.Goal.t -> query
 val query_of_term :
-  program -> (depth:int -> State.t -> State.t * (Loc.t * Elpi_data.term)) -> query * Elpi_data.term StrMap.t
+  program -> (depth:int -> State.t -> State.t * (Loc.t * term)) -> query
 
-val pp_query : (depth:int -> Format.formatter -> Elpi_data.term -> unit) -> Format.formatter -> query -> unit
+val pp_query : (depth:int -> Format.formatter -> term -> unit) -> Format.formatter -> query -> unit
 
 val executable_of_query : query -> Elpi_data.executable
 
-val term_of_ast : depth:int -> Loc.t * Elpi_ast.Term.t -> Elpi_data.term
+val term_of_ast : depth:int -> Loc.t * Elpi_ast.Term.t -> term
 
-type quotation = depth:int -> State.t -> Loc.t -> string -> State.t * Elpi_data.term
+type quotation = depth:int -> State.t -> Loc.t -> string -> State.t * term
 val set_default_quotation : quotation -> unit
 val register_named_quotation : name:string -> quotation -> unit
 
 val lp : quotation
 
-val is_Arg : State.t -> Elpi_data.term -> bool
-val fresh_Arg : 
-  State.t -> name_hint:string -> args:Elpi_data.term list ->
-    State.t * string * Elpi_data.term
+val is_Arg : State.t -> term -> bool
+val get_Args : State.t -> term StrMap.t
+val mk_Arg : 
+  State.t -> name:string -> args:term list ->
+    State.t * term
 
 (* Quotes the program and the query, see elpi_quoted_syntax.elpi *)
-val quote_syntax : query -> Elpi_data.term list * Elpi_data.term
+val quote_syntax : query -> term list * term
 
 (* false means a type error was found *)
 val static_check : Elpi_ast.Program.t -> (* header *)
@@ -46,3 +48,5 @@ val static_check : Elpi_ast.Program.t -> (* header *)
   ?checker:Elpi_ast.Program.t ->
   ?flags:flags ->
   query -> bool
+
+val while_compiling : bool State.component